Extract body editor.
